Create "careful" draw and erase routines
These careful methods ensure that only bits set in the shape table
are erased. This is needed because the padding of each shape, which
is necessary for frame shifting, causes bits from adjacent or
overlapping shapes to be erased.

Note that the erasing routine is just another entry point to the
drawing routine, allowing it to use most all of the same code.

The non-careful erasing routine has been renamed to "erase_square,"
which is a more accurate name since no shape is passed to it.