For draw and erase, set bit 7 differently for different situations
This fixes problems when the snake and an apple share the same byte,
and we have to decide which color to use. In addition to fixing
graphical glitches, it fixes a serious problem when eating apples
horizontally. Often, the apple turns green, and then the game ends
because the game thinks the snake collided with itself!

It is still not perfect. It is possible to change an apple green when
eating an apple next to it, and there are color glitches when eating,
but the game now seems to be playable!